[map-functions] Move titleCase out of context
