+++ packages/map/src/components/CityList.tsx
@@ -5,6 +5,7 @@ import { supportedCities } from '../data/supported-geos'
 import { scaleLinear } from 'd3-scale'
 import { GlyphStar, GlyphTriangle, GlyphDiamond, GlyphSquare, GlyphCircle } from '@visx/glyph'
 import { getFilterControllingStatePicked } from './UsaMap/helpers/map'
+import { titleCase } from '../helpers/titleCase'
 
 import ConfigContext from '../context'
 import { getGeoStrokeColor } from '../helpers/colors'
@@ -15,14 +16,13 @@ const CityList = ({
   applyTooltipsToGeo,
   displayGeoName,
   applyLegendToRow,
-  titleCase,
   setSharedFilterValue,
   isFilterValueSupported,
   tooltipId,
   projection
 }) => {
   const [citiesData, setCitiesData] = useState({})
-  const { scale, state, topoData, runtimeData, position } = useContext(ConfigContext)
+  const { state, topoData, runtimeData, position } = useContext(ConfigContext)
   if (!projection) return
 
   useEffect(() => {
