+++ packages/map/src/components/WorldMap/WorldMap.tsx
@@ -14,6 +14,7 @@ import ZoomControls from '../ZoomControls'
 import { getGeoFillColor, getGeoStrokeColor } from '../../helpers/colors'
 import { supportedCountries } from '../../data/supported-geos'
 import { handleMapAriaLabels } from '../../helpers/handleMapAriaLabels'
+import { titleCase } from '../../helpers/titleCase'
 
 const { features: world } = feature(topoJSON, topoJSON.objects.countries)
 
@@ -35,10 +36,7 @@ const WorldMap = () => {
     setRuntimeData,
     setState,
     state,
-    titleCase,
     tooltipId,
-    setScale,
-    setTranslate
   } = useContext(ConfigContext)
 
   // TODO Refactor - state should be set together here to avoid rerenders
