+++ packages/map/src/context.ts
@@ -32,7 +32,6 @@ type MapContext = {
   setSharedFilterValue
   setState
   state: MapConfig
-  titleCase
   tooltipId: string
   viewport
 }
